tests/wpt/web-platform-tests/fullscreen/api/element-request-fullscreen-timing-manual.html

<!DOCTYPE html>
<title>Element#requestFullscreen() timing</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<div id="log"></div>
<script>
async_test(t => {
  const div = document.querySelector('div');
  trusted_request(t, div);

  // If fullscreenchange is an animation frame event, then animation frame
  // callbacks should be run after it is fired, before the timer callback.
  // The resize event should fire before the fullscreenchange event.
  const events = [];
  const callback = t.step_func(event => {
    // fullscreenElement should have changed before either event is fired.
    assert_equals(document.fullscreenElement, div, `fullscreenElement in {event.type} event`);
    events.push(event.type);
    if (event.type == 'fullscreenchange') {
      step_timeout(t.unreached_func('timer callback'));
      requestAnimationFrame(t.step_func_done(() => {
        assert_array_equals(events, ['resize', 'fullscreenchange'], 'event order');
      }));
    }
  });
  document.onfullscreenchange = window.onresize = callback;
}, 'Timing of fullscreenchange and resize events');

async_test(t => {
  var promise = document.createElement('a').requestFullscreen();
  var promise_executed = false;
  if (promise) {
    promise.catch(()=>{promise_executed = true; });
  } else {
    // if promises aren't supported treat it as executed.
    promise_executed = true;
  }

  // If fullscreenerror is an animation frame event, then animation frame
  // callbacks should be run after it is fired, before the timer callback.
  document.onfullscreenerror = t.step_func(() => {
    assert_true(promise_executed, "promise executed");
    step_timeout(t.unreached_func('timer callback'));
    requestAnimationFrame(t.step_func_done());
  });
}, 'Timing of fullscreenerror event');
</script>
